src/mte_tag: sw control eliding tag checks on per-pointer basis
Even though page table encoding allows software to control whether a page
is tagged or not, software may need to avoid tag checks even though page
is tagged. Such situations are memory accesses to local objects to a
function. Only when pointer to local objects is passed to another
function, it is required to be checked for tags. Thus providing a control
in spec for eliding tag checks on per-pointer basis using a special bypass
`pointer_tag` value of 0.

Signed-off-by: Deepak Gupta <debug@rivosinc.com>