src/mte_tag: `settag`, `gentag` and `addtag` are independent of MTAG=1
MTAG=1 for codepage only impacts generation of checked loads and stores.
Behavior of `settag`, `gentag` and `addtag` instructions are not impacted.
This is primarily for the reason that user must have not compiled with this
option. If they compiled it with this option then all the memory objects
created by this region of code should be creating tagged pointers and setting
metadata correctly (even though code itself is subject waiver of checking
tags).

Signed-off-by: Deepak Gupta <debug@rivosinc.com>