src/mte_tag: tighten the first stage page table conditions for MTAG (#79)
Quite often loader or other piece of code might read code pages. If a code
page is set as MTAG=1, then those loads (to code page) will look for tags.
This change clarifies in specification that loads/stores to pages marked
as executable always generate unchecked memory accesses.

If if memory tagging is not enabled for execution environment then `MTAG`
bit is ignored. This allows for setting up tagging and enabling/disabling
on sampling basis (without asking OS to clear to the bit when tagging is
disabled)

Signed-off-by: Deepak Gupta <debug@rivosinc.com>